server/requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2

# Data validation và settings
pydantic>=2.3.0
pydantic-settings>=2.1.0
python-dotenv>=0.19.0
email-validator>=2.0.0

# Database
supabase>=2.3.0
postgrest>=0.13.0
nest-asyncio==1.5.8

# Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt==4.0.1
cryptography>=45.0.2

# HTTP và Networking
httpx>=0.23.0

# Testing
pytest==7.3.1

# Websockets
websockets==11.0.3

# Additional package
certifi>=2024.2.2

# New dependencies
tenacity==8.2.3
cachetools==5.3.2
prometheus-client>=0.11.0,<0.12.0

# Added from the code block
aiofiles>=0.7.0,<0.8.0
python-magic>=0.4.24,<0.5.0
pillow>=10.0.0
boto3>=1.18.0,<1.19.0
aioredis==2.0.1

# Added from the code block
APScheduler==3.10.4

# Timezone handling
pytz>=2024.1

# Added from the code block
web3==7.12.0
mnemonic>=0.21
eth-account>=0.13.7
bip-utils>=2.9.3

requests>=2.32.3
tweepy
sqlalchemy